+++ src/extension.type.ts
@@ -1,0 +1,12 @@
+export type Model = Record<string, any>;
+
+export interface FlatModel {
+  name: string;
+  value: Model;
+}
+
+export enum NotificationCode {
+  interfaceNameInvalid = 'interfaceNameInvalid',
+  interfaceNameContainSpaces = 'interfaceNameContainSpaces',
+  notValidObject = 'notValidObject',
+}