+++ src/interface-builder.types.ts
@@ -1,9 +1,0 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-export type Model = Record<string, any>;
-
-export interface FlatModel {
-  name: string;
-  value: Model;
-}
-
-export type OptionKeys = 'addabove' | 'replace' | 'newfile';