+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.0
+Added replacing selected object by ts interface
+Added export ts interface to file
+Added configuration default interface name
+Added configuration export file postfix
+Added validation entered interface name
+Added notification
 # 0.0.6
 Fix union type
 # 0.0.5